delete and restore routes add,destroy route update

+++ routes/web.php
@@ -96,9 +96,12 @@ Route::prefix('user/')->middleware('auth')->name('user.')->group(function () {
     Route::resource('prospects', ProspectsController::class)->only(['create', 'store'])->middleware('permissions:prospect-write-web');
     Route::get('prospects/{prospect}', [ProspectsController::class, 'show'])->middleware('permissions:prospect-read-web')->name('prospects.show');
     Route::resource('prospects', ProspectsController::class)->only(['edit', 'update'])->middleware('permissions:prospect-edit-web');
-    
-    Route::delete('prospects/{prospect}', [ProspectsController::class, 'destroy'])->name('prospects.destroy');
-    
+
+    Route::delete('prospects/{prospect}', [ProspectsController::class, 'delete'])->middleware('permissions:prospect-update-web')->name('prospects.delete');
+    Route::put('prospects/{prospect}/restore', [ProspectsController::class, 'restore'])->middleware('permissions:prospect-update-web')->name('prospects.restore');
+
+    Route::delete('prospects/{prospect}/force_delete', [ProspectsController::class, 'destroy'])->middleware('permissions:prospect-delete-web')->name('prospects.destroy');
+
     Route::get('prospects/{prospect}/orders', [ProspectsController::class, 'show'])->middleware('permissions:order-read-web')->name('prospects.show-orders')->where('prospect', '[0-9]+');
 
     //Products
